requirements.txt
streamlit
requests
newspaper3k
transformers
sentence-transformers
faiss-cpu
langchain
openai
sqlalchemy
pandas
python-dotenv
tqdm